+++ docs/doc/changes.rst
@@ -43,7 +43,6 @@ Rather we work on something else?  Just let us know!
 
 Backlog:
 
-  * |in_progress| Task: Replace edit alias with edit or create label rule.
   * Task: Flush browsers site cache when a new version of application is used.
   * Task: Provide a link to the RPi's astutus-web-app.
   * Task: Have an indication that publish wheels is proceding.
@@ -71,6 +70,7 @@ hand waving, but not as detailed as repository commits.
 
 Done:
 
+  * |newly_done| 2021-02-05 14:00 Task: Replace edit alias with edit or create label rule.
   * |newly_done| 2021-02-05 11:14 Task: Refactor to use DeviceClassifier labels instead of aliases.
   * |newly_done| 2021-02-05 11:14 Task: Remove dead alias code.
   * |newly_done| 2021-02-05 10:43 Task: Remove last of DeviceConfiguration code.